Components: Adds ping system support to target portrait
The target portrait is a large area of the target unit frame that
might be intuitively used by players to ping their target.

While oUF adds this by itself to the actual unit frame, the overlay
frame added by AzeriteUI will not inherit the functionality.

This change adds the PingableTypeMixin to the overlay on retail WoW.