Components: Add addon memory usage tooltip to framerate widget
The default Blizzard interface displays framerate, latency, and
addon memory usage in the tooltip for the main menu button in the
MicroMenu.

AzeriteUI displays framerate and latency in its info widget, but has
no equivalent display for addon memory usage.

This change adds a tooltip to the fps widget with that information,
using similar code to Blizzard's own tooltip code.